Create separate FieldType and MethodType classes
