+++ elide-core/src/main/java/com/yahoo/elide/core/type/ClassType.java
@@ -6,18 +6,15 @@
 
 package com.yahoo.elide.core.type;
 
-import org.apache.commons.lang3.reflect.TypeUtils;
 import lombok.Getter;
 
 import java.lang.annotation.Annotation;
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.ParameterizedType;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Date;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 public class ClassType<T> implements Type<T> {
@@ -168,88 +165,7 @@ public class ClassType<T> implements Type<T> {
             return null;
         }
 
-        return new Field() {
-
-            @Override
-            public void setAccessible(boolean flag) {
-                field.setAccessible(flag);
-            }
-
-            @Override
-            public boolean isSynthetic() {
-                return field.isSynthetic();
-            }
-
-            @Override
-            public boolean isAnnotationPresent(Class<? extends Annotation> annotationClass) {
-                return field.isAnnotationPresent(annotationClass);
-            }
-
-            @Override
-            public <T extends Annotation> T getAnnotation(Class<T> annotationClass) {
-                return field.getAnnotation(annotationClass);
-            }
-
-            @Override
-            public <T extends Annotation> T[] getAnnotationsByType(Class<T> annotationClass) {
-                return field.getAnnotationsByType(annotationClass);
-            }
-
-            @Override
-            public Annotation[] getDeclaredAnnotations() {
-                return field.getDeclaredAnnotations();
-            }
-
-            @Override
-            public Annotation[] getAnnotations() {
-                return field.getAnnotations();
-            }
-
-            @Override
-            public int getModifiers() {
-                return field.getModifiers();
-            }
-
-            @Override
-            public String getName() {
-                return field.getName();
-            }
-
-            @Override
-            public Object get(Object obj) throws IllegalArgumentException, IllegalAccessException {
-                return field.get(obj);
-            }
-
-            @Override
-            public Type<?> getType() {
-                return new ClassType(field.getType());
-            }
-
-            @Override
-            public Type<?> getParameterizedType(Type<?> parentType, Optional<Integer> index) {
-                if (parentType instanceof DynamicType) {
-                    return getType();
-                }
-
-                java.lang.reflect.Type type = field.getGenericType();
-
-                if (type instanceof ParameterizedType && index.isPresent()) {
-                    return new ClassType(
-                            TypeUtils.getRawType(
-                                ((ParameterizedType) type).getActualTypeArguments()[index.get().intValue()],
-                                ((ClassType) parentType).getCls()
-                            )
-                    );
-                }
-                return new ClassType(TypeUtils.getRawType(type, ((ClassType) parentType).getCls()));
-            }
-
-            @Override
-            public void set(Object obj, Object value) throws IllegalArgumentException, IllegalAccessException {
-                field.set(obj, value);
-            }
-
-        };
+        return new FieldType(field);
     }
 
     public static Method constructMethod(java.lang.reflect.Executable method) {
@@ -257,102 +173,7 @@ public class ClassType<T> implements Type<T> {
             return null;
         }
 
-        return new Method() {
-            @Override
-            public int getModifiers() {
-                return method.getModifiers();
-            }
-
-            @Override
-            public void setAccessible(boolean flag) {
-                method.setAccessible(flag);
-            }
-
-            @Override
-            public boolean isSynthetic() {
-                return method.isSynthetic();
-            }
-
-            @Override
-            public boolean isAnnotationPresent(Class<? extends Annotation> annotation) {
-                return method.isAnnotationPresent(annotation);
-            }
-
-            @Override
-            public <T extends Annotation> T getAnnotation(Class<T> annotationClass) {
-                return method.getAnnotation(annotationClass);
-            }
-
-            @Override
-            public <T extends Annotation> T[] getAnnotationsByType(Class<T> annotationClass) {
-                return method.getAnnotationsByType(annotationClass);
-            }
-
-            @Override
-            public Annotation[] getDeclaredAnnotations() {
-                return method.getDeclaredAnnotations();
-            }
-
-            @Override
-            public Annotation[] getAnnotations() {
-                return method.getAnnotations();
-            }
-
-            @Override
-            public String getName() {
-                return method.getName();
-            }
-
-            @Override
-            public int getParameterCount() {
-                return method.getParameterCount();
-            }
-
-            @Override
-            public Object invoke(Object obj, Object... args) throws IllegalAccessException,
-                    IllegalArgumentException, InvocationTargetException {
-                if (! (method instanceof java.lang.reflect.Method)) {
-                    throw new UnsupportedOperationException("Constructors cannot be invoked");
-                }
-                return ((java.lang.reflect.Method) method).invoke(obj, args);
-            }
-
-            @Override
-            public Type<?> getReturnType() {
-                if (! (method instanceof java.lang.reflect.Method)) {
-                    throw new UnsupportedOperationException("Constructors cannot be invoked");
-                }
-                return new ClassType(((java.lang.reflect.Method) method).getReturnType());
-            }
-
-            @Override
-            public Type<?> getParameterizedReturnType(Type<?> parentType, Optional<Integer> index) {
-                if (! (method instanceof java.lang.reflect.Method)) {
-                    throw new UnsupportedOperationException("Constructors cannot be invoked");
-                }
-
-                if (parentType instanceof DynamicType) {
-                    return getReturnType();
-                }
-
-                java.lang.reflect.Type type = ((java.lang.reflect.Method) method).getGenericReturnType();
-
-                if (type instanceof ParameterizedType && index.isPresent()) {
-                    return new ClassType(
-                            TypeUtils.getRawType(
-                                    ((ParameterizedType) type).getActualTypeArguments()[index.get().intValue()],
-                                    ((ClassType) parentType).getCls()
-                            )
-                    );
-                }
-                return new ClassType(TypeUtils.getRawType(type, ((ClassType) parentType).getCls()));
-            }
-
-            @Override
-            public Class<?>[] getParameterTypes() {
-                return method.getParameterTypes();
-            }
-        };
+        return new MethodType(method);
     }
 
     public static Package constructPackage(java.lang.Package pkg) {
